Added tests for nesting, equalities and for un-named VisualizationFrames
